[MAT-317] factor out Fake_{JVM,JNIEnv} to header file for easy enclusion and deduping code.
