+++ include/test/fake_jvm.hh
@@ -1,0 +1,71 @@
+/**
+ * Copyright (C) 2014 - present by OpenGamma Inc. and the OpenGamma group of companies
+ *
+ * Please see distribution for license.
+ */
+
+#include "jvmmanager.hh"
+#include "warningmacros.h"
+
+//
+// Contains concrete impls of fake jvm and fake jni env
+//
+
+// mock JVM
+class Fake_JavaVM: public JavaVM
+{
+  public:
+    virtual ~Fake_JavaVM(){};
+    virtual int GetEnv(void ** env, int SUPPRESS_UNUSED version) override
+    {
+     *env = _env;
+     return JNI_OK;
+    }
+    void setEnv(JNIEnv * env)
+    {
+        this->_env = env;
+    }
+    JNIEnv * getEnv()
+    {
+        return this->_env;
+    }
+    virtual jint AttachCurrentThread(void **penv, void SUPPRESS_UNUSED *args)
+    {
+      *penv = this->_env;
+      return JNI_OK;
+    }
+};
+
+// mock JNIEnv
+class Fake_JNIEnv: public JNIEnv
+{
+  public:
+    Fake_JNIEnv()
+    {
+      _someNonClass = new _jclass();
+      _someNonMethodID = new _jmethodID();
+      _someNonFieldID = new _jfieldID();
+    }
+    virtual ~Fake_JNIEnv()
+    {
+      delete _someNonMethodID;
+      delete _someNonFieldID;
+      delete _someNonClass;
+    }
+    virtual jclass FindClass(const char SUPPRESS_UNUSED * name)
+    {
+      return _someNonClass;
+    }
+    virtual jmethodID GetMethodID(jclass SUPPRESS_UNUSED clazz, const char SUPPRESS_UNUSED *name, const char SUPPRESS_UNUSED *sig) override
+    {
+      return _someNonMethodID;
+    }
+    virtual jfieldID GetFieldID(jclass SUPPRESS_UNUSED clazz, const char SUPPRESS_UNUSED *name, const char SUPPRESS_UNUSED *sig) override
+    {
+      return _someNonFieldID;
+    }
+private:
+    jclass _someNonClass;
+    jmethodID _someNonMethodID;
+    jfieldID _someNonFieldID;
+};
