+++ src/jshim/test/check_jbindings.cc
@@ -13,68 +13,10 @@
 #include <iostream>
 #include "jvmmanager.hh"
 #include "debug.h"
+#include "test/fake_jvm.hh"
 
 using namespace std;
 using namespace convert;
-
-// mock JVM
-class Fake_JavaVM: public JavaVM
-{
-  public:
-    virtual ~Fake_JavaVM(){};
-    virtual int GetEnv(void ** env, int SUPPRESS_UNUSED version) override
-    {
-     *env = _env;
-     return JNI_OK;
-    }
-    void setEnv(JNIEnv * env)
-    {
-        this->_env = env;
-    }
-    JNIEnv * getEnv()
-    {
-        return this->_env;
-    }
-    virtual jint AttachCurrentThread(void **penv, void SUPPRESS_UNUSED *args)
-    {
-      *penv = this->_env;
-      return JNI_OK;
-    }
-};
-
-// mock JNIEnv
-class Fake_JNIEnv: public JNIEnv
-{
-  public:
-    Fake_JNIEnv()
-    {
-      _someNonClass = new _jclass();
-      _someNonMethodID = new _jmethodID();
-      _someNonFieldID = new _jfieldID();
-    }
-    virtual ~Fake_JNIEnv()
-    {
-      delete _someNonMethodID;
-      delete _someNonFieldID;
-      delete _someNonClass;
-    }
-    virtual jclass FindClass(const char SUPPRESS_UNUSED * name)
-    {
-      return _someNonClass;
-    }
-    virtual jmethodID GetMethodID(jclass SUPPRESS_UNUSED clazz, const char SUPPRESS_UNUSED *name, const char SUPPRESS_UNUSED *sig) override
-    {
-      return _someNonMethodID;
-    }
-    virtual jfieldID GetFieldID(jclass SUPPRESS_UNUSED clazz, const char SUPPRESS_UNUSED *name, const char SUPPRESS_UNUSED *sig) override
-    {
-      return _someNonFieldID;
-    }
-private:
-    jclass _someNonClass;
-    jmethodID _someNonMethodID;
-    jfieldID _someNonFieldID;
-};
 
 class bindRunner
 {
