+++ src/jshim/test/check_jterminals.cc
@@ -14,73 +14,10 @@
 #include "jvmmanager.hh"
 #include "debug.h"
 #include "equals.hh"
+#include "test/fake_jvm.hh"
 
 using namespace std;
 using namespace convert;
-
-
-// mock JVM
-class Fake_JavaVM: public JavaVM
-{
-  public:
-    virtual ~Fake_JavaVM(){};
-    virtual int GetEnv(void ** env, int SUPPRESS_UNUSED version) override
-    {
-     *env = _env;
-     return JNI_OK;
-    }
-    void setEnv(JNIEnv * env)
-    {
-        this->_env = env;
-    }
-    JNIEnv * getEnv()
-    {
-        return this->_env;
-    }
-    virtual jint AttachCurrentThread(void SUPPRESS_UNUSED **penv, void SUPPRESS_UNUSED *args)
-    {
-      *penv = this->_env;
-      return JNI_OK;
-    }
-};
-
-// mock JNIEnv
-class Fake_JNIEnv: public JNIEnv
-{
-  public:
-    Fake_JNIEnv()
-    {
-      _someNonClass = new _jclass();
-      _someNonMethodID = new _jmethodID();
-      _someNonFieldID = new _jfieldID();
-    }
-    virtual ~Fake_JNIEnv()
-    {
-      delete _someNonMethodID;
-      delete _someNonFieldID;
-      delete _someNonClass;
-    }
-    virtual jclass FindClass(const char SUPPRESS_UNUSED * name)
-    {
-      return _someNonClass;
-    }
-    virtual jmethodID GetMethodID(jclass SUPPRESS_UNUSED clazz, const char SUPPRESS_UNUSED *name, const char SUPPRESS_UNUSED *sig) override
-    {
-      return _someNonMethodID;
-    }
-    virtual jfieldID GetFieldID(jclass SUPPRESS_UNUSED clazz, const char SUPPRESS_UNUSED *name, const char SUPPRESS_UNUSED *sig) override
-    {
-      return _someNonFieldID;
-    }
-    using JNIEnv::CallIntMethod;
-    using JNIEnv::CallObjectMethod;
-    using JNIEnv::GetIntArrayElements;
-    using JNIEnv::GetDoubleArrayElements;
-private:
-    jclass _someNonClass;
-    jmethodID _someNonMethodID;
-    jfieldID _someNonFieldID;
-};
 
 
 TEST(JTerminals, Test_getIntFromVoidJMethod_null_meth)
